+++ ROCKMAN2/Rockman2/scene/SceneTitle.cpp
@@ -163,11 +163,11 @@ void SceneTitle::Draw()
 	DrawBg();
 	
 	// ロゴ表示
-	DrawRotaGraph(Game::kScreenWidth * 0.5, Game::kScreenHeight * 0.25 - 40, 1.0f, 0.0f, m_logoHandle, true);
+	DrawRotaGraph(static_cast<int>(Game::kScreenWidth * 0.5), static_cast<int>(Game::kScreenHeight * 0.25 - 40), 1.0f, 0.0f, m_logoHandle, true);
 	// 文字表示
 	DrawRectRotaGraph(kCharPosX, kCharPosY, 0, 0, kSelectSizeX, kSelectSizeY, 1.0f, 0.0f, m_charHandle, true, false);
 	// 選択カーソルの表示
-	DrawRectRotaGraph(m_selectPos.x, m_selectPos.y, 0, 0, kSelectSizeX, kSelectSizeY, 1.0f, 0.0f, m_selectHandle, true, false);
+	DrawRectRotaGraph(static_cast<int>(m_selectPos.x), static_cast<int>(m_selectPos.y), 0, 0, kSelectSizeX, kSelectSizeY, 1.0f, 0.0f, m_selectHandle, true, false);
 
 	// フェード描画
 	SetDrawBlendMode(DX_BLENDMODE_ALPHA, m_fadeAlpha);
@@ -199,8 +199,8 @@ void SceneTitle::DrawBg()
 	for (int index = 0; index < 2; index++)
 	{
 		DrawRotaGraph2(
-			scrollBg2 + index * bg2Size.width * kBgScale,
-			Game::kScreenHeight - bg2Size.height * kBgScale,
+			static_cast<int>(scrollBg2 + index * bg2Size.width * kBgScale),
+			static_cast<int>(Game::kScreenHeight - bg2Size.height * kBgScale),
 			0, 0,
 			kBgScale, 0.0f,
 			m_bg2Handle, true);
@@ -209,8 +209,8 @@ void SceneTitle::DrawBg()
 	for (int index = 0; index < 2; index++)
 	{
 		DrawRotaGraph2(
-			scrollBg3 + index * bg3Size.width * kBgScale,
-			Game::kScreenHeight - bg3Size.height * kBgScale,
+			static_cast<int>(scrollBg3 + index * bg3Size.width * kBgScale),
+			static_cast<int>(Game::kScreenHeight - bg3Size.height * kBgScale),
 			0, 0,
 			kBgScale, 0.0f,
 			m_bg3Handle, true);
@@ -219,8 +219,8 @@ void SceneTitle::DrawBg()
 	for (int index = 0; index < 2; index++)
 	{
 		DrawRotaGraph2(
-			scrollBg4 + index * bg4Size.width * kBgScale,
-			Game::kScreenHeight - bg4Size.height * kBgScale,
+			static_cast<int>(scrollBg4 + index * bg4Size.width * kBgScale),
+			static_cast<int>(Game::kScreenHeight - bg4Size.height * kBgScale),
 			0, 0,
 			kBgScale, 0.0f,
 			m_bg4Handle, true);
